Habit Tracker Web Application/requirements.txt
celery==5.2.7
email-validator==1.2.1
Flask==2.2.2
Flask-Caching==2.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-RESTful==0.3.9
Flask-Security-too
Flask-SQLAlchemy==2.5.1
jsonpickle==2.2.0
matplotlib==3.5.3
PyJWT==2.4.0
redis==4.3.4
SQLAlchemy==1.4.40
weasyprint==56.1
Werkzeug==2.2.2